User mgmt drop in component (#71)
* SLA-1941 Create layout components and fetch account settings

* SLA-1944 Create edit name for account settings

* SLA-1943 Create summary page and navigation

* Create a way to delete connections and change account settings in slashauth core client

* Revert removal of close signin modal when clicking outside

Although this seemed like a good options at first. The sign-in modal is
also used to connect wallets, email and other to your existing account.
There is no way for the user to escape once started this flow, which is
a terrible ux.

* Fix openSignInSync: promise will resolve when opened more than once

When an event that can be subscribed was triggered inside the
componentController, the listeners are notified. However the reference
to the list of listeners was never updated, even when a listener was
removed and a new one was added.

This caused the openSingInSync to only resolve once. Which caused the
consumer of openSignInSync to not get notified when the user opened and
closed the sign in modal more than once.

This was extremely painful in the user mgmt component where we need to
fetch the user account settings after connecting to a wallet or other
login method.

* Create an avatar component to be reused in the dropdown component

* Create chip component

* Display web3 login methods even if the user is already signed with one

* SLA-1942 User profile image editing in account settings

* Hide scroll bar on signin component

* 1.4.0